This change is needed to correctly mount the internal storage (internal SD card) on Android Kitkat.
It should be applied to the ramdisk used in the boot.img file.
To port it to other platforms, simply replace "yukon" and "flamingo"
in file and directory names with whatever device you are using.

The problem is twofold: wrong SELinux permissions on the
/data/media/0 mount point and a failure to start the FUSE sdcard
service from the init.rc script.

As a result /mnt/shell/emulated is empty, instead of bringing the
content of /data/media (including the internal SD card /data/media/0).

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>